[feat] move the dlock from sc to eventbase
